fix price datatype

+++ database/migrations/2018_03_22_194858_create_product_table.php
@@ -17,7 +17,7 @@ class CreateProductTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('description');
-            $table->integer('price');
+            $table->double('price', 8, 2);
             $table->timestamps();
         });
     }
